Fix notification LEDs, add amber LED and update defconfig.
Add options for how LEDs are handled when system suspends. I may add an option to blink later, but the cost on battery may be unacceptable.
Add amber LED. This is basically working but is a bit buggy if blinking at the same time as other leds.
Update config.

Conflicts:

	arch/arm/configs/cyanogen_blade_defconfig
	drivers/leds/leds-msm-pmic-status.c